requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
oauthlib==3.2.2
requests==2.31.0
werkzeug==2.3.7
Pillow==10.2.0
psycopg2-binary==2.9.9
Flask-Migrate==4.1.0

